storage/util: move GetCurrentResourceVersionFromStorage
